Fixing declaration/implementation mismatch
I'm not entirely sure whether the declaration is wrong or the implementation, so if I got this backwards, apologies. But, there is presently no implementation of `gb_float33_9` and no declaration of `gb_float33_16`, so I assumed one was supposed to match the other. (I'm not actually using this function, but discovered the missing part when doing some automatic Lua bindings based on importing headers.)